requirements-airflow.txt
apache-airflow==2.8.0
apache-airflow-providers-email==1.2.0
python-dateutil>=2.8.2

